fix: closure bugs and range int64 support
- Fix closure variable detection for keyword argument values (func(typed=typed) pattern)
- Fix class body closure handling (function → class → method pattern)
- Change PyRange from int to long for larger range support
- Add PyType support in ABCMeta.register()
- Add _typing.py stub module for future CPython typing.py compatibility

+++ Parser/ast.cs
@@ -1841,7 +1841,7 @@ namespace SharpPy
                 }
                 else if (iterable is PyRange range)
                 {
-                    for (int i = range.Start; i < range.Stop; i += range.Step)
+                    for (long i = range.Start; i < range.Stop; i += range.Step)
                     {
                         AssignTarget(scope, Target, new PyInt(i));
                         try
